@@ -8,13 +8,12 @@
   },
   "homepage": "http://react-bootstrap.github.io/",
   "scripts": {
-    "build": "./node_modules/.bin/grunt build",
+    "build": "./run-babel tools/build.js",
     "test-watch": "karma start",
     "test": "karma start --single-run && npm run lint",
-    "babel-build-amd": "babel --modules amd --optional es7.objectRestSpread src --out-dir tmp-amd",
-    "babel-build-lib": "babel --optional es7.objectRestSpread src --out-dir tmp-lib",
-    "lint": "eslint src test docs ie8 webpack karma.conf.js webpack.config.js webpack.docs.js",
-    "build-docs": "webpack --config webpack.docs.js && node docs/build.js"
+    "lint": "eslint src test docs ie8 tools webpack karma.conf.js webpack.config.js webpack.docs.js",
+    "docs-build": "./run-babel tools/build.js --docs-only",
+    "docs-run": "./run-babel docs/server.js"
   },
   "main": "lib/main.js",
   "directories": {
@@ -32,52 +31,33 @@
     "react": ">=0.12"
   },
   "devDependencies": {
-    "async": "~0.2.9",
     "babel": "^4.7.0",
     "babel-core": "^4.7.4",
-    "babel-eslint": "^1.0.13",
+    "babel-eslint": "^2.0.2",
     "babel-loader": "^4.1.0",
     "bootstrap": "3.2.0",
     "brfs": "^1.4.0",
     "child-process-promise": "^1.0.1",
     "client-loader": "0.0.1",
     "codemirror": "^5.0.0",
+    "colors": "^1.0.3",
     "css-loader": "^0.9.1",
-    "envify": "~1.2.1",
     "es5-shim": "^4.1.0",
-    "eslint": "^0.15.1",
-    "eslint-plugin-react": "^1.4.1",
+    "eslint": "^0.17.1",
+    "eslint-plugin-react": "^1.6.0",
     "express": "^3.4.8",
     "extract-text-webpack-plugin": "^0.3.8",
     "file-loader": "^0.8.1",
     "fs-promise": "^0.3.1",
-    "grunt": "~0.4.2",
-    "grunt-amd-wrap": "^1.0.1",
-    "grunt-browserify": "~1.3.0",
-    "grunt-cli": "~0.1.13",
-    "grunt-contrib-clean": "~0.5.0",
-    "grunt-contrib-copy": "~0.5.0",
-    "grunt-contrib-requirejs": "~0.4.1",
-    "grunt-contrib-uglify": "~0.3.2",
-    "grunt-contrib-watch": "~0.5.3",
-    "grunt-karma": "^0.10.1",
-    "grunt-react": "~0.10.0",
-    "grunt-shell": "~0.6.4",
     "json-loader": "^0.5.1",
-    "karma": "~0.12.8",
-    "karma-babel-preprocessor": "^4.0.0",
+    "karma": "~0.12.32",
     "karma-chai": "0.0.2",
     "karma-chrome-launcher": "~0.1.2",
     "karma-cli": "0.0.4",
-    "karma-coffee-preprocessor": "~0.1.2",
     "karma-firefox-launcher": "~0.1.3",
-    "karma-html2js-preprocessor": "~0.1.0",
-    "karma-jasmine": "~0.1.5",
     "karma-mocha": "~0.1.1",
     "karma-mocha-reporter": "^1.0.2",
     "karma-phantomjs-launcher": "~0.1.1",
-    "karma-requirejs": "~0.2.1",
-    "karma-script-launcher": "~0.1.0",
     "karma-sinon": "^1.0.3",
     "karma-sourcemap-loader": "^0.3.4",
     "karma-webpack": "^1.5.0",
@@ -86,9 +66,7 @@
     "lodash": "^3.5.0",
     "mocha": "~1.16.2",
     "react": "~0.12.0",
-    "react-async": "~2.0.0",
     "react-router-component": "https://github.com/STRML/react-router-component/archive/v0.23.0.tar.gz",
-    "requirejs": "~2.1.9",
     "rf-changelog": "^0.4.0",
     "semver": "~2.0.7",
     "sinon": "^1.10.3",
