@@ -1,27 +1,23 @@
-'use strict';
+import express from 'express';
+import path from 'path';
+import url from 'url';
+import webpack from 'webpack';
+import webpackMiddleware from 'webpack-dev-middleware';
+import webpackConfigBuilder from '../webpack/webpack.config';
+import Root from './src/Root';
 
-var express = require('express');
-
-var development = process.env.NODE_ENV !== 'production';
-var app = express();
-var path = require('path');
+const development = process.env.NODE_ENV !== 'production';
+let app = express();
 
 if (development) {
-  require('babel/register');
-  var path = require('path');
-  var url = require('url');
-  var webpack = require('webpack');
-  var webpackMiddleware = require('webpack-dev-middleware');
-  var webpackConfig = require('../webpack/webpack.config')({
+  let webpackConfig = webpackConfigBuilder({
     development: development,
     docs: true
   });
-  var publicPath = webpackConfig.output.publicPath;
+  let publicPath = webpackConfig.output.publicPath;
 
   webpackConfig.output.path = '/';
   webpackConfig.output.publicPath = undefined;
-
-  var Root = require('./src/Root');
 
   app = app
     .use(webpackMiddleware(webpack(webpackConfig), {
@@ -32,9 +28,8 @@ if (development) {
       }
     }))
     .use(function renderApp(req, res) {
-      var fileName = url.parse(req.url).pathname;
-      var RootHTML = Root.renderToString({initialPath: fileName});
-
+      let fileName = url.parse(req.url).pathname;
+      let RootHTML = Root.renderToString({initialPath: fileName});
       res.send(RootHTML);
     });
 } else {
