@@ -1,46 +1,46 @@
 /* eslint-disable */
 import React from 'react';
 import classSet from 'react/lib/cx';
-import Accordion from '../../lib/Accordion';
-import Alert from '../../lib/Alert';
-import Badge from '../../lib/Badge';
-import Button from '../../lib/Button';
-import ButtonGroup from '../../lib/ButtonGroup';
-import ButtonToolbar from '../../lib/ButtonToolbar';
-import CollapsableMixin from '../../lib/CollapsableMixin';
-import Carousel from '../../lib/Carousel';
-import CarouselItem from '../../lib/CarouselItem';
-import Col from '../../lib/Col';
-import DropdownButton from '../../lib/DropdownButton';
-import Glyphicon from '../../lib/Glyphicon';
-import Grid from '../../lib/Grid';
-import Input from '../../lib/Input';
-import Jumbotron from '../../lib/Jumbotron';
-import Label from '../../lib/Label';
-import ListGroup from '../../lib/ListGroup';
-import ListGroupItem from '../../lib/ListGroupItem';
-import Nav from '../../lib/Nav';
-import Navbar from '../../lib/Navbar';
-import NavItem from '../../lib/NavItem';
-import MenuItem from '../../lib/MenuItem';
-import Modal from '../../lib/Modal';
-import ModalTrigger from '../../lib/ModalTrigger';
-import OverlayTrigger from '../../lib/OverlayTrigger';
-import OverlayMixin from '../../lib/OverlayMixin';
-import PageHeader from '../../lib/PageHeader';
-import PageItem from '../../lib/PageItem';
-import Pager from '../../lib/Pager';
-import Panel from '../../lib/Panel';
-import PanelGroup from '../../lib/PanelGroup';
-import Popover from '../../lib/Popover';
-import ProgressBar from '../../lib/ProgressBar';
-import Row from '../../lib/Row';
-import SplitButton from '../../lib/SplitButton';
-import TabbedArea from '../../lib/TabbedArea';
-import Table from '../../lib/Table';
-import TabPane from '../../lib/TabPane';
-import Tooltip from '../../lib/Tooltip';
-import Well from '../../lib/Well';
+import Accordion from '../../src/Accordion';
+import Alert from '../../src/Alert';
+import Badge from '../../src/Badge';
+import Button from '../../src/Button';
+import ButtonGroup from '../../src/ButtonGroup';
+import ButtonToolbar from '../../src/ButtonToolbar';
+import CollapsableMixin from '../../src/CollapsableMixin';
+import Carousel from '../../src/Carousel';
+import CarouselItem from '../../src/CarouselItem';
+import Col from '../../src/Col';
+import DropdownButton from '../../src/DropdownButton';
+import Glyphicon from '../../src/Glyphicon';
+import Grid from '../../src/Grid';
+import Input from '../../src/Input';
+import Jumbotron from '../../src/Jumbotron';
+import Label from '../../src/Label';
+import ListGroup from '../../src/ListGroup';
+import ListGroupItem from '../../src/ListGroupItem';
+import Nav from '../../src/Nav';
+import Navbar from '../../src/Navbar';
+import NavItem from '../../src/NavItem';
+import MenuItem from '../../src/MenuItem';
+import Modal from '../../src/Modal';
+import ModalTrigger from '../../src/ModalTrigger';
+import OverlayTrigger from '../../src/OverlayTrigger';
+import OverlayMixin from '../../src/OverlayMixin';
+import PageHeader from '../../src/PageHeader';
+import PageItem from '../../src/PageItem';
+import Pager from '../../src/Pager';
+import Panel from '../../src/Panel';
+import PanelGroup from '../../src/PanelGroup';
+import Popover from '../../src/Popover';
+import ProgressBar from '../../src/ProgressBar';
+import Row from '../../src/Row';
+import SplitButton from '../../src/SplitButton';
+import TabbedArea from '../../src/TabbedArea';
+import Table from '../../src/Table';
+import TabPane from '../../src/TabPane';
+import Tooltip from '../../src/Tooltip';
+import Well from '../../src/Well';
 /* eslint-enable */
 
 import {CodeMirror, IS_NODE} from './CodeMirror';
