@@ -1,5 +1,5 @@
 /* eslint no-var: 0 */
-require('babel/register');
+require('./register-babel');
 var config = require('./webpack/webpack.config');
 
 var result = config();
