@@ -1,3 +1,3 @@
-require('babel/register');
+require('./register-babel');
 var config = require('./webpack/docs.config');
 module.exports = config;
